add obe detail interface and obe detail api for lecturer

+++ course-service/course-service/src/main/java/hcmuaf/edu/vn/fit/course_service/repository/EnrollmentRepository.java
@@ -16,7 +16,7 @@ public interface EnrollmentRepository extends JpaRepository<Enrollment, String> 
 
     // 1. Sửa lại thành studentId (Vì trong Enrollment đổi thành String studentId)
     List<Enrollment> findByStudentId(String studentId);
-    long countByCourseOffering_OfferingId(String offeringId);
+    int countByCourseOffering_OfferingId(String offeringId);
     // 2. Sửa lại thành studentId
     Optional<Enrollment> findByStudentIdAndCourseOffering_OfferingId(String studentId, String offeringId);
 
@@ -45,4 +45,5 @@ public interface EnrollmentRepository extends JpaRepository<Enrollment, String> 
             "AND e.status = 'ACTIVE' COLLATE utf8mb4_unicode_520_ci",
             nativeQuery = true)
     List<StudentCourseProjection> findStudentsByOfferingId(@Param("offeringId") String offeringId);
+
 }
